% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilp-setup.R
\name{all_connections}
\alias{all_connections}
\title{Find all connections of an element}
\usage{
all_connections(x, y)
}
\arguments{
\item{i}{The point for which the connections are sought (must be in 1...n)}

\item{n}{The number of points}
}
\value{
A `data.frame` with two columns representing start and end point.
  The second column always contains the "larger number" point.
}
\description{
Find all connections of an element
}
