% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anticlust-package.R
\docType{package}
\name{anticlust-package}
\alias{anticlust}
\alias{anticlust-package}
\title{\packageTitle{anticlust}}
\description{
\packageDescription{anticlust}
}
\details{
The DESCRIPTION file:
\tabular{ll}{
Package: \tab anticlust\cr
Type: \tab Package\cr
Version: \tab 0.3.0-1\cr
Date: \tab 2019-11-11\cr
Depends: \tab R (>= 3.4.0)\cr
Encoding: \tab UTF-8\cr
License: \tab MIT + file LICENSE\cr
URL: \tab https://github.com/m-Py/anticlust\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/m-Py/anticlust}
  \item Report bugs at \url{https://github.com/m-Py/anticlust/issues}
}

}
\author{
\packageAuthor{anticlust}

Maintainer: \packageMaintainer{anticlust}
}
