% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicriterion_iterated_local_search_call.R
\name{bicriterion_iterated_local_search_call}
\alias{bicriterion_iterated_local_search_call}
\title{Solve anticlustering by using the BILS algorithm by Brusco et al.}
\usage{
bicriterion_iterated_local_search_call(
  data,
  G,
  R,
  W = c(1e-06, 1e-05, 1e-04, 0.001, 0.01, 0.1, 0.5, 0.99, 0.999, 0.999999),
  Xi = c(0.05, 0.1)
)
}
\arguments{
\item{data}{A N x N dissimilarity matrix or N x M features matrix.}

\item{G}{The number of clusters}

\item{R}{The desired number of restarts for the algorithm}

\item{W}{Optional argument, a vector of possible weights for dispersion 
in the bicriterion (0 <= W <= 1)}

\item{Xi}{Optional argument, a vector for the neighborhood size range
[xi1, xi2] in percent (for example c(0.05, 0.1) can be interpreted as 
5%-10%)}
}
\value{
List of useful partitions (approximated pareto set)
}
\description{
Solve anticlustering by using the BILS algorithm by Brusco et al.
}
