% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnb.R
\name{bnb_anticlustering}
\alias{bnb_anticlustering}
\title{Simple branch and bound approach to anticlustering}
\usage{
bnb_anticlustering(features, K)
}
\arguments{
\item{features}{A vector, matrix or data.frame of data points. Rows
correspond to elements and columns correspond to features. A
vector represents a single feature.}

\item{K}{How many anticlusters should be created.}
}
\value{
A list containing the objective and the anticluster
    affiliations.
}
\description{
Simple branch and bound approach to anticlustering
}
