% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anticlustering_Spaeth.R
\name{cluster_centers}
\alias{cluster_centers}
\title{Compute cluster centers}
\usage{
cluster_centers(features, clusters)
}
\arguments{
\item{features}{A data matrix of element features}

\item{clusters}{A numeric vector indicating cluster membership
of each element}
}
\value{
A matrix of cluster centers. Rows represent clusters and
  columns represent features
}
\description{
Compute cluster centers
}
