% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilp-setup.R
\name{connections_to_pair}
\alias{connections_to_pair}
\title{Get character representation of item connections}
\usage{
connections_to_pair(connections)
}
\arguments{
\item{connections}{a data.frame returned by `all_connections`}
}
\value{
A character vector representing item connections.
}
\description{
Get character representation of item connections
}
