% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted-cluster-editing.R
\name{cpp_similarities}
\alias{cpp_similarities}
\title{Compute similarity values for the clique partitioning problem}
\usage{
cpp_similarities(x)
}
\arguments{
\item{x}{A N x M data frame / matrix of attributes}
}
\value{
An object of class dist
}
\description{
Compute similarity values for the clique partitioning problem
}
