% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective-anticluster-editing.R
\name{distance_objective}
\alias{distance_objective}
\title{Objective value for the cluster editing distance objective}
\usage{
distance_objective(x, clusters)
}
\arguments{
\item{x}{The data input. Can be one of two structures: (1) A data matrix
where rows correspond to elements and columns correspond to
features (a single numeric feature can be passed as a vector). (2)
An N x N matrix dissimilarity matrix; can be an object of class
\code{dist} (e.g., returned by \code{\link{dist}} or
\code{\link{as.dist}}) or a \code{matrix} where the entries of
the upper and lower triangular matrix represent the pairwise
dissimilarities.}

\item{clusters}{A vector representing (anti)clusters (e.g.,
returned by \code{\link{anticlustering}} or
\code{\link{balanced_clustering}}).}
}
\value{
The cluster editing objective
}
\description{
Check the objective value for a given clustering.
}
\details{
The cluster editing objective objective is given by the sum of the
pairwise distances between elements within the same (anti)clusters.
When the argument \code{features} is passed, the Euclidean distance
is used.
}
\examples{

data(iris)
distances <- dist(iris[1:60, -5])
## Clustering
clusters <- balanced_clustering(distances, K = 3)
# This is low:
distance_objective(distances, clusters)
## Anticlustering
anticlusters <- anticlustering(distances, K = 3)
# This is higher:
distance_objective(distances, anticlusters)


# Illustrates the cluster editing objective as the sum of distances
# within groups (needs an integer linear programming solver!)
n_elements <- 12
features <- matrix(runif(n_elements * 2), ncol = 2)
n_groups <- 2
clusters <- balanced_clustering(features, K = n_groups, method = "ilp")
anticlusters <- anticlustering(features, K = n_groups, method = "ilp")
par(mfrow = c(1, 2))
plot_clusters(features, clusters, within_connection = TRUE,
              main = "Minimum within-group distances")
plot_clusters(features, anticlusters, within_connection = TRUE,
              main = "Maximum within-group distances")

}
\references{
Papenberg, M., & Klau, G. W. (2019, October 30). Using anticlustering
to partition a stimulus pool into equivalent parts.
https://doi.org/10.31234/osf.io/3razc
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
