% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_select.R
\name{divide_and_select}
\alias{divide_and_select}
\title{Divide and Select}
\usage{
divide_and_select(split_by, equalize, design, n, p)
}
\arguments{
\item{split_by}{The features by which the data is split}

\item{design}{Specifies the number of levels per \code{split_by} 
feature. Is a vector of length \code{ncol(split_by)} (or of length
1 if only one \code{split_by} feature is passed.}

\item{n}{The number of elements per set.}

\item{p}{The number of exchange partners; higher values increase
the precision of the results but also increase run time.}

\item{select}{The features that should be similar across sets}
}
\value{
The anticluster affiliation
}
\description{
Stimulus selection through the divide and select approach. 
Makes some variables dissimilar between sets, other variables
similar between sets.
}
\examples{
# TODO

TODO control parameters: 
  - number of elements per group (as matrix?)
  - p can be control parameter
  - thresholds on divide parameter
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
