% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{draw_assignment}
\alias{draw_assignment}
\title{Visualize a clustering of two features}
\usage{
draw_assignment(features, clustering, xlab = "feature 1",
  ylab = "feature 2", cols = NULL, pch = 19, main = "", cex = 3,
  connect_clusters = FALSE)
}
\arguments{
\item{features}{A data.frame or matrix representing the features that
are used. Must have two columns.}

\item{clustering}{A vector representing the clustering}

\item{xlab}{The label for the x-axis}

\item{ylab}{The label for the y-axis}

\item{cols}{The coloring of the groups. Does not need to be passed.
If it is passed, it needs to be a character vector of the same
length as there are groups, and each element of the vector is a
color.}

\item{pch}{The symbol used for data points, see ?par}

\item{main}{The title of the plot}

\item{cex}{The size of the plotting symbols, see ?par}

\item{connect_clusters}{Boolean. Connect the groups through lines?
Useful to illustrate a graph structure.}
}
\description{
Visualize a clustering of two features
}
