% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heuristic.R
\name{edit_distances}
\alias{edit_distances}
\title{Edit distances of very close neighbours (similar items)}
\usage{
edit_distances(distances, assignment, value = -1e+06)
}
\arguments{
\item{distances}{A distance object or matrix of
between-item-distances.}

\item{value}{The value that is assigned to the fixed distances.
Defaults to -1,000,000 currently.}
}
\value{
A vector representing the group assignment; objects that
  are part of the same group as indicated by this vector are assigned
  a new distance.

A distance object containing the fixed distances. For items
    that are part of the same cluster (as specified in the
    `data.frame` `assignment`), the between-item-distances are set to
    -1000000. This will have to be replaced by a theoretically-sound
    value; -1000000 is just a hack that will work in the present
    applications.
}
\description{
Based on a preclustering, distances between items of the same cluster
are set to a large negative value. By considering a preclustering of
items, very similar items will not be assigned to the same group when
the fixed distance object is used to create the ILP formulation of
the item assignment instance.
}
