% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete-enumeration.R
\name{enum_anticlustering}
\alias{enum_anticlustering}
\title{Complete enumeration approach to anticluster editing}
\usage{
enum_anticlustering(features, K)
}
\arguments{
\item{features}{A vector, matrix or data.frame of data points. Rows
correspond to elements and columns correspond to features. A
vector represents a single feature.}

\item{K}{How many anticlusters should be created.}
}
\value{
The anticluster affiliation
}
\description{
Complete enumeration approach to anticluster editing
}
\examples{

n_features <- 2
N <- 10
K <- 2
features <- matrix(runif(n_features * N), ncol = n_features)
results <- enum_anticlustering(features, K, objective = "distance")
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
