% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilp-solve.R
\name{equal_sized_cluster_editing}
\alias{equal_sized_cluster_editing}
\title{Solve exact equal-sized cluster editing}
\usage{
equal_sized_cluster_editing(features, n_clusters, solver,
  standardize = FALSE)
}
\arguments{
\item{features}{A vector, matrix or data.frame of data points.  Rows
correspond to items and columns correspond to features.}

\item{n_clusters}{How many clusters are to be created.}

\item{solver}{A string identifing the solver to be used ("Rglpk",
"gurobi", or "Rcplex")}

\item{standardize}{Boolean - should the feature values be
standardized before groups are created? Defaults to FALSE.}
}
\value{
A vector representing the clustering.
}
\description{
Solve exact equal-sized cluster editing
}
