% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heuristic.R
\name{equal_sized_clustering}
\alias{equal_sized_clustering}
\title{Heuristic clustering algorithm creating equal-sized clusters}
\usage{
equal_sized_clustering(items, nclusters)
}
\arguments{
\item{items}{A vector, matrix or data.frame of data points. If a
matrix or data.frame is passed, rows correspond to items and
columns correspond to features.}

\item{nclusters}{The number of clusters to be created. Must be a
factor of the number of items.}
}
\value{
A data.frame containing one column of item ids (here, the id
    corresponds to the order of the items) and one column contains the
    group assignments of the items. The original items are also
    returned as columns of the data.frame.
}
\description{
Uses kmeans algorithm to initiate cluster centers and then
sequentially chooses a cluster center and assigns assigns the closest
item to it. This ensures that each cluster is filled with the same
number of items.
}
