% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilp-setup.R
\name{equality_identifiers}
\alias{equality_identifiers}
\title{Based on the solver, return identifiers for equality relationships}
\usage{
equality_identifiers(solver)
}
\arguments{
\item{solver}{A string identifing the solver to be used ("Rglpk",
"gurobi", or "cplex")}
}
\value{
A list of three elements containing strings representing
    equality (e), lower (l), and greater (g) relationships
}
\description{
Based on the solver, return identifiers for equality relationships
}
