% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c-anticlustering.R
\name{fanticlust}
\alias{fanticlust}
\title{C implementation of anticlustering}
\usage{
fanticlust(data, K, categories = NULL)
}
\arguments{
\item{data}{An N x M data matrix.}

\item{K}{The number of clusters or an initial assignment of the elements 
to clusters.}

\item{categories}{A vector, data.frame or matrix representing one
or several categorical constraints.}
}
\description{
C implementation of anticlustering
}
\examples{
# Compare classical anticlustering implementation and C implementation
N <- 100
M <- 3
data <- matrix(rnorm(N * M), ncol = M)
K <- 3
C <- 3
categories <- sample(C, size = N, replace = TRUE)
start <- Sys.time()
cl1 <- fanticlust(data, K, categories)
Sys.time() - start
table(categories, cl1)

start <- Sys.time()
cl2 <- anticlustering(
  data,
  K = clusters,
  objective = "variance"
)
Sys.time() - start

all(cl1 == cl2) # Same algorithm in C and R


}
