% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/wrapper-feasible-and-infeasible-region-search.R
\name{feasible_and_infeasible_region_search_anticlustering}
\alias{feasible_and_infeasible_region_search_anticlustering}
\title{Feasible and Infeasible Region search heuristic}
\usage{
feasible_and_infeasible_region_search_anticlustering(
  x,
  K,
  N,
  objective = "diversity",
  number_iterations = 50,
  clusters = NULL,
  upper_bound = NULL,
  lower_bound = NULL,
  beta_max = NULL,
  theta_max = NULL,
  theta_min = NULL,
  beta_min = NULL,
  phi = 0.8,
  tau = NULL,
  kmax = 4,
  alpha = NULL
)
}
\arguments{
\item{K}{Number of anticlusters to be formed.}

\item{N}{Number of elememts.}

\item{objective}{The anticlustering objective = "diversity".}

\item{clusters}{A vector of length M that specifies the number of elements each cluster can contain. 
If this vector is not NULL, the lower and upper bounds will be disregarded.}

\item{upper_bound}{Maximum number of elements in each anticluster. By default, anticlusters are of equal size,
calculated as the total number of items divided by the number of clusters.}

\item{lower_bound}{Minimum number of elements in each anticluster. By default, anticlusters are of equal size,
calculated as the total number of items divided by the number of clusters.}

\item{beta_max}{The algorithm begins with a pool of random initial solutions of size beta_max. 
Over time, the size of the solution pool decreases linearly until it reaches beta_min.}

\item{theta_max}{Parameter for the strength of undirected perturbation, 
which decreases linearly over time from theta_max to theta_min.}

\item{theta_min}{Parameter for the strength of undirected perturbation, 
which decreases linearly over time from theta_max to theta_min.}

\item{beta_min}{The minimum solution pool size the algorithm should reach before making a determination.}

\item{phi}{Parameter that determines the population size when initiating a new round of exploration upon
triggering the jump-back mechanism}

\item{tau}{Parameter when "noImp" exceeds it, the algorithm transits from exploitation to exploration strategy}

\item{alpha}{Parameter for weitghing the discrimitation of a slighlty worse local optiomal child solution.}

\item{matrix}{The data input. Currently just a vector.}

\item{number_iterations:}{A number that defines how many times the steps in the search algorithm are repeated.}

\item{kmax:}{A parameter that determines the maximum degree of constraint violation in the IFR search}

\item{return}{results contains everything, including vector result and its cost}
}
\value{
A vector of length N that assigns a group (i.e, a number
    between 1 and \code{K}) to each input element.
}
\description{
This function implements the feasible and infeasible region search algorithm FIFR for
anticlustering by Wu et al. (2025; <doi.org/10.1016/j.cor.2025.107030>).
The description of their algorithm is given in Section 4 of their paper (in particular, see the
Pseudocode in Algorithm 1).
}
\details{
Details of the implementation of the algorithm can be found 
 in the pseudocode of the paper Wu et al. (2025)
}
\examples{

# Generate some random data
N <- 120
M <- 5
K <- 3
dat <- matrix(rnorm(N * M), ncol = M)
distances <- dist(dat)

# Perform three hase serach algorithm
results1 <- anticlust:::feasible_and_infeasible_region_search_anticlustering(distances, K, N, number_iterations = 50)
results2 <- anticlustering(distances, K = K, method = "3phase")
results3 <- anticlustering(distances, K = K, method = "local-maximum", repetitions = 50)

# Compute objectives funtion
diversity_objective(distances, results1)
diversity_objective(distances, results2)
diversity_objective(distances, results3)


}
\references{
Wu, X., Feng, J., Yang, J., & Zhang, Y. (2025). Feasible and infeasible region 
search for the maximally diverse grouping problem. Computers & Operations Research, 179,
107030. https://doi.org/10.1016/j.cor.2025.107030
}
\author{
David Buczynski \email{david.buczynski@hhu.de}, 
Martin Papenberg \email{martin.papenberg@hhu.de}
}
