% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-exchange-partners.R
\name{generate_exchange_partners}
\alias{generate_exchange_partners}
\title{Generate exchange partners for anticlustering}
\usage{
generate_exchange_partners(N = NULL, p, categories = NULL,
  similar = FALSE, features = NULL)
}
\arguments{
\item{N}{The number of elements. Does not need to be passed if either
\code{categories}, \code{features} or \code{distance} is passed.}

\item{p}{The number of exchange partners per element.}

\item{categories}{A vector describing categories of elements.}

\item{similar}{Logical. Should similar items serve as exchange partners?
Defaults to \code{FALSE}.}

\item{features}{Optional data argument. If the exchange partners
are generated via a cluster analysis (that is if argument
\code{similar = TRUE}), data is needed describing the elements.}
}
\description{
This function is used to generate a vector that is used as the
\code{categories} argument in the function \code{\link{anticlustering}}.
This function is meant to reduce the total number of exchange partners
(the default is *all* elements) so that the exchange method runs
faster.
}
\details{
This function will be particularly interesting when generating exchange
partners within categories. For imbalanced data (that is, within each
category, the elements cannot be split into equal-sized parts of
size \code{p}, the function may return some elements with \code{p + 1}
exchange partners.

In general, the function will try to balance out everything as well as
possible (this is also true for clustering and especially clustering
within categories), but it cannot always succeed. That is, sometimes
the function will throw an error if the argument \code{p} does not
fit with the arguments \code{similar = TRUE} (which request a cluster
analysis with clusters of equal size) or \code{categories}.
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
