% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast-exchange-method.R
\name{get_neigbours}
\alias{get_neigbours}
\title{Get neigbours for fast preclustering (by category)}
\usage{
get_neigbours(features, k_neighbours, categories)
}
\description{
Get neigbours for fast preclustering (by category)
}
