% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-objective.R
\name{get_objective}
\alias{get_objective}
\title{Compute objective value for the (anti)clustering problem}
\usage{
get_objective(features, anticlusters, objective)
}
\arguments{
\item{features}{A data.frame, matrix or vector representing the
features that are used in (anti)clustering.}

\item{anticlusters}{A vector representing the anticluster affiliation}

\item{objective}{The objective to be maximized, either "distance" or
"variance".}
}
\value{
Scalar, the objective value; either the total
  within-cluster variance (if objective == "variance") or
  the total sum of within-cluster pointwise distances
  (if objective == "distance")
}
\description{
Compute objective value for the (anti)clustering problem
}
\examples{

features <- matrix(rnorm(1000, 100, 15), ncol = 2)
n_anticlusters <- 4
# Precluster cases
n_preclusters <- nrow(features) / n_anticlusters
preclusters <- equal_sized_kmeans(features, n_preclusters)
# Use preclustering as resticting information in anticlustering
anticlusters <- heuristic_anticlustering(features, preclusters, objective = "distance")
get_objective(features, anticlusters, "distance")
get_objective(features, anticlusters, "variance")
anticlusters <- heuristic_anticlustering(features, preclusters, objective = "variance")
get_objective(features, anticlusters, "distance")
get_objective(features, anticlusters, "variance")

}
