% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy-matching.R
\name{greedy_matching}
\alias{greedy_matching}
\title{Greedy matching}
\usage{
greedy_matching(distances)
}
\arguments{
\item{distances}{A N x N matrix representing the
pairwise dissimilarities between N elements. Larger values
indicate higher dissimilarity. Can be an object of class
\code{dist} (e.g., returned by \code{\link{dist}} or
\code{\link{as.dist}}).}
}
\value{
A vector representing which elements were matched.
}
\description{
Finds pairs of elements that are similar. Can be used as a
preclustering step for heuristic anticluster editing when K = 2.
}
\details{
This function realizes a greedy matching approach given pairwise
dissimilarities of N elements. The algorithm works as follows:

1. Determine the two closest elements with regard to the
   pairwise dissimilarity

2. Assign these two elements to the same cluster

3. Remove these two elements from the pool

4. Repeat 1-3 until no elements are remaining
}
\examples{

# Compare exact and greedy matching
N <- 10
n_features <- 2
features <- matrix(rnorm(N * n_features), ncol = n_features)
c1 <- greedy_matching(dist(features))
c2 <- clustering(features, K = N / 2, method = "exact")
par(mfrow = c(1, 2))
plot_clusters(features, c1, within_connection = TRUE, cex = 2.5)
plot_clusters(features, c2, within_connection = TRUE, cex = 2.5)

}
