% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md.R
\name{mahalanobis_distance}
\alias{mahalanobis_distance}
\title{Compute mahalanobis distance between two vectors}
\usage{
mahalanobis_distance(x, y, cov_mat, lambda = 0)
}
\arguments{
\item{x}{First vector}

\item{y}{First vector}

\item{cov_mat}{covariance matrix between the features in x and y}

\item{lambda}{Regularization, defaults to 0 (no regularization)}
}
\value{
The mahalanobis distance (scalar value)
}
\description{
Compute mahalanobis distance between two vectors
}
\examples{
mahalanobis_distance(iris[1, -5], iris[2, -5], cov(iris[, -5]))

}
