% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_sd_objective-function.R
\name{mean_sd_obj}
\alias{mean_sd_obj}
\title{An objective function measuring similarity of sets}
\usage{
mean_sd_obj(clusters, features)
}
\arguments{
\item{clusters}{A clustering vector}

\item{features}{The features}
}
\value{
A value quantifying set similarity as the total absolute deviations
    between sets on all features. Higher values indicate that the deviations
    in mean, median and standard deviations are low.
}
\description{
Uses variance in means, medians, and standard deviations per feature
across sets as the objective.
}
