% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives_by_cluster.R
\name{mean_sd_tab}
\alias{mean_sd_tab}
\title{Means and standard deviations by group variable formatted in table}
\usage{
mean_sd_tab(features, groups, decimals = 2, na.rm = FALSE, return_diff = FALSE)
}
\arguments{
\item{features}{A data frame of features}

\item{groups}{A grouping vector}

\item{decimals}{The number of decimals}

\item{na.rm}{Should NAs be removed prior to computing stats 
(Default = FALSE)}

\item{return_diff}{Boolean. Should an additional row be printed that 
contains the difference between minimum and maximum}
}
\value{
A table that illustrates means and standard deviations (in brackets)
}
\description{
Means and standard deviations by group variable formatted in table
}
\examples{

data(iris)
mean_sd_tab(iris[, -5], iris[, 5])

}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
