% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/must-link-anticlustering.R
\name{must_link_anticlustering}
\alias{must_link_anticlustering}
\title{Must-link anticlustering}
\usage{
must_link_anticlustering(x, K, must_link, method = "exchange")
}
\arguments{
\item{x}{A distance matrix (Only maximum diversity is implemented)}

\item{K}{What do the required groups look like (as in \code{\link{anticlustering}})}

\item{must_link}{Vector of IDs; elements having the same value are linked in the same anticluster.}

\item{method}{"exchange" (default) or "local-maximum"}
}
\description{
Must-link anticlustering
}
\details{
Conducts an exchange method on "merged" elements.
}
