% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number-of-partitions.R
\name{n_partitions}
\alias{n_partitions}
\title{Number of equal sized partitions}
\usage{
n_partitions(N, K)
}
\arguments{
\item{N}{How many elements}

\item{K}{How many partitions}
}
\value{
The number of partitions
}
\description{
Number of equal sized partitions
}
\examples{
n_partitions(20, 2)

}
