% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next-permutation.R
\name{next_permutation}
\alias{next_permutation}
\title{Get the next permutation}
\usage{
next_permutation(permutation)
}
\arguments{
\item{permutation}{A vector of elements.}
}
\value{
The next higher permutation of the elements in vector
    `permutation` with regard to its lexicographic ordering.
}
\description{
Each permutation is computed on basis of a passed permutation where
lexicographic ordering of the permutations is used to determine the
next "higher" permutation. This is an adaption of the
`next_permutation` function in C++.
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
