% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heuristic.R
\name{obj_value}
\alias{obj_value}
\title{Get the objective function from an assignment object}
\usage{
obj_value(assignment)
}
\arguments{
\item{assignment}{A data.frame that represents a clustering
assingment. Usually, the return value of `ilp_to_groups` or
`equal_sized_clustering` will be the input.}
}
\value{
The objective value
}
\description{
Get the objective function from an assignment object
}
