% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-objective.R
\name{obj_value_distance}
\alias{obj_value_distance}
\title{Objective value for the distance criterion}
\usage{
obj_value_distance(features, anticlusters)
}
\arguments{
\item{features}{A data.frame, matrix or vector representing the
features that are used in the assignment.}

\item{anticlusters}{A vector representing the anticluster affiliation}
}
\value{
Scalar, the total sum of within-cluster
  pointwise distances.
}
\description{
Objective value for the distance criterion
}
