% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variance_Criterion.R
\name{obj_value_variance}
\alias{obj_value_variance}
\title{Compute objective value for variance criterion}
\usage{
obj_value_variance(features, anticlusters)
}
\arguments{
\item{features}{A data.frame, matrix or vector representing the
features that are used in the assignment.}

\item{anticlusters}{A vector representing the anticluster affiliation}
}
\value{
Scalar: the objective value
}
\description{
Compute objective value for variance criterion
}
