% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Visualize a clustering of two features}
\usage{
plot_clusters(features, clustering, connect_clusters = FALSE,
  xlab = NULL, ylab = NULL, cols = NULL, pch = 19, main = "",
  cex = 1.2, cex.axis = 1.2, cex.lab = 1.2)
}
\arguments{
\item{features}{A data.frame or matrix representing the features that
are used. Must have two columns.}

\item{clustering}{A vector representing the clustering}

\item{connect_clusters}{Boolean. Connect the groups through lines?
Useful to illustrate a graph structure.}

\item{xlab}{The label for the x-axis}

\item{ylab}{The label for the y-axis}

\item{cols}{The coloring of the groups. Does not need to be passed.
If it is passed, it needs to be a character vector of the same
length as there are groups, and each element of the vector is a
color.}

\item{pch}{The symbol used for data points, see ?par}

\item{main}{The title of the plot}

\item{cex}{The size of the plotting symbols, see ?par}

\item{cex.axis}{The size of the values on the axes}

\item{cex.axis}{The size of the labels of the axes}
}
\description{
Visualize a clustering of two features
}
