% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-test-set.R
\name{plot_test_clusters}
\alias{plot_test_clusters}
\title{Plot Clusters of Data Points}
\usage{
plot_test_clusters(
  data,
  labels,
  N_clusters = 7,
  N_data_per_cluster = 6,
  std = 0.2,
  save_file = FALSE
)
}
\arguments{
\item{data}{A matrix of shape 2 x (N_clusters * N_data_per_cluster), where each column 
represents the coordinates (X, Y) of a data point.}

\item{labels}{A numeric vector of length (N_clusters * N_data_per_cluster) containing 
the cluster labels for each data point.}

\item{N_clusters}{Integer. The number of clusters to plot. Default is 7.}

\item{N_data_per_cluster}{Integer. The number of data points per cluster. Default is 6.}

\item{std}{Double. The standard deviation for generating cluster points,
controlling the spread of the clusters. Default is 0.2.}

\item{save_file}{Boolean. Saves the clustering image to a file.}
}
\value{
This function does not return any value. It creates a scatter plot.
}
\description{
This function creates a 2D scatter plot of data points colored by cluster.
It is intended to visualize the results of the `generate_test_clusters` function.
}
\examples{
result <- generate_test_clusters()
plot_test_clusters(result$data, result$labels)

}
