% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-sort.R
\name{sort_by_col}
\alias{sort_by_col}
\title{Utility function to sort a matrix or data frame by one column}
\usage{
sort_by_col(x, col)
}
\arguments{
\item{x}{A matrix or data frame}

\item{row}{An index for row selection (numeric; character; logical)}
}
\value{
The matrix or data frame sorted by `row`
}
\description{
Utility function to sort a matrix or data frame by one column
}
\examples{
mat <- cbind(10:1, 1:10)
mat
sort_by_col(mat, 1)
}
