% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilp-setup.R
\name{sparse_constraints}
\alias{sparse_constraints}
\title{Construct a sparse matrix representing the ILP constraints}
\usage{
sparse_constraints(n_items, pair_names)
}
\arguments{
\item{n_items}{How many items are there}

\item{pair_names}{A character vector of names representing the item
pairs.  Must have the form that is contained in the ILP
data.frame costs$pair.}
}
\value{
A sparse matrix representing the left-hand side of the ILP (A
    in Ax ~ b)
}
\description{
Construct a sparse matrix representing the ILP constraints
}
