% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical-variable-handling.R
\name{to_numeric}
\alias{to_numeric}
\title{Converting any categorical input to integer}
\usage{
to_numeric(x)
}
\arguments{
\item{x}{a vector}
}
\value{
A vector of \code{length(x)} where each element is integer
}
\description{
Converting any categorical input to integer
}
\examples{
iris$Species
to_numeric(iris$Species)

}
