% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective-functions.R
\name{variance_objective}
\alias{variance_objective}
\title{Objective value for the variance criterion}
\usage{
variance_objective(features, clusters, standardize = FALSE)
}
\arguments{
\item{features}{A vector, matrix or data.frame of data points. Rows
correspond to elements and columns correspond to features. A
vector represents a single feature.}

\item{clusters}{A vector representing (anti)clusters (e.g., returned
by \code{\link{anticlustering}} or
\code{\link{balanced_clustering}})}

\item{standardize}{Boolean - should the features be standardized
before the objective is computed? Defaults to \code{FALSE}.
Standardization is done using the function \code{\link{scale}}
using the default settings (mean = 0, SD = 1).}
}
\value{
The total within-cluster variance
}
\description{
Check the objective value for a given clustering.
}
\details{
The variance objective is given by the sum of the squared
errors between cluster centers and individual data points. It is the
objective function used in k-means clustering, see
\code{\link{kmeans}}.
}
\note{
When using this function to check the results of
\code{\link{anticlustering}} or \code{\link{balanced_clustering}},
make sure that the \code{standardization} argument has the same
value when creating (anti)clusters and when calling
\code{variance_objective}.
}
\examples{

data(iris)
## Clustering
clusters <- balanced_clustering(
  iris[, -5],
  K = 3,
  standardize = TRUE,
  objective = "variance"
)
# This is low:
variance_objective(
  iris[, -5],
  clusters,
  standardize = TRUE
)
## Anticlustering
anticlusters <- anticlustering(
  iris[, -5],
  K = 3,
  standardize = TRUE,
  objective = "variance",
  nrep = 100
)
# This is higher:
variance_objective(
  iris[, -5],
  anticlusters,
  standardize = TRUE
)

}
\references{
Jain, A. K. (2010). Data clustering: 50 years beyond k-means.
Pattern Recognition Letters, 31, 651–666.

Späth, H. (1986). Anticlustering: Maximizing the variance criterion.
Control and Cybernetics, 15, 213–218.
}
