% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILP.R
\name{vectorize_weights}
\alias{vectorize_weights}
\title{Convert matrix of distances into vector of distances}
\usage{
vectorize_weights(distances)
}
\arguments{
\item{distances}{A distance matrix}
}
\value{
A data.frame having the following columns:
   `costs` - the actual weights in vectorized form
   `i` the first index of the item pair that is connected
   `j` the second index of the item pair that is connected
   `pair` A string of form x_ij identifying the item pair
}
\description{
Convert matrix of distances into vector of distances
}
