% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilp-setup.R
\name{vectorized_group}
\alias{vectorized_group}
\title{Construct indices for a sparse matrix representation of group
constraints}
\usage{
vectorized_group(n_items, pair_names)
}
\arguments{
\item{n_items}{How many items does the instance have}

\item{pair_names}{A character vector of names representing the item
pairs.  Must have the form that is contained in the ILP
data.frame costs$pair.}
}
\value{
A list of indices to be used as input parameters of
    Matrix::sparseMatrix
}
\description{
Construct indices for a sparse matrix representation of group
constraints
}
