% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anticlustering.R
\name{anticlustering}
\alias{anticlustering}
\title{Anticlustering}
\usage{
anticlustering(features, n_anticlusters, standardize = TRUE,
  objective = "distance", method = "heuristic")
}
\arguments{
\item{features}{A vector, matrix or data.frame of data points.  Rows
correspond to elements and columns correspond to features.}

\item{n_anticlusters}{How many anticlusters should be created.}

\item{standardize}{Boolean - should the features be standardized
before anticlusters are created? Defaults to TRUE}

\item{objective}{The objective to be maximized, either "distance"
(default) or "variance". See Details}

\item{method}{One of "heuristic" or "exact". Currently, an exact
solution can only be obtained when the `objective` is "distance".}
}
\value{
A vector representing anticluster affiliation
}
\description{
Create groups of elements (anticlusters) that are as similar as
possible.
}
\details{
TODO: Write details
}
\examples{

# TODO: Write examples

}
