% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-wrapper.R
\name{balanced_clustering}
\alias{balanced_clustering}
\title{Create K balanced clusters}
\usage{
balanced_clustering(features, K, standardize = TRUE,
  method = "heuristic")
}
\arguments{
\item{features}{A vector, matrix or data.frame of data points. Rows
correspond to elements and columns correspond to features.}

\item{K}{How many clusters should be created.}

\item{standardize}{Boolean - should the features be standardized
before anticlusters are assigned? Defaults to TRUE}

\item{method}{One of "heuristic" or "exact". When "exact" is chosen,
the function minimizes the distance criterion (i.e., solves
balanced k-cluster editing); when "heuristic"
is chosen, the function tries to minimize the variance criterion
(see \code{\link{anticlustering}} for details). Use
\code{method = "exact"} only for rather small n.}
}
\value{
A vector representing the cluster affiliation of all elements.
    Each cluster has the the same size.
}
\description{
Create K balanced clusters
}
\examples{

data(iris)
# Only use numeric attributes
clusters <- balanced_clustering(iris[, -5], K = 3)
# Compare feature means by anticluster
by(iris[, -5], clusters, function(x) round(colMeans(x), 2))
# Plot the anticlustering
par(mfrow = c(1, 2))
plot_clusters(iris[, 1:2], clusters)
plot_clusters(iris[, 3:4], clusters)
par(mfrow = c(1, 1))

## Exact balanced cluster editing method
# Create artifical data
n_features <- 2
n_elements <- 30
features <- matrix(rnorm(n_elements * n_features), ncol = n_features)
ac <- balanced_clustering(features, K = 2, method = "exact")
plot_clusters(features, ac, within_connection = TRUE)

}
