% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories_to_binary.R
\name{categories_to_binary}
\alias{categories_to_binary}
\title{Get binary representation of categorical variables}
\usage{
categories_to_binary(categories, use_combinations = FALSE)
}
\arguments{
\item{categories}{A vector, data.frame or matrix representing one
or several categorical variables}

\item{use_combinations}{Logical, should the output also include columns representing
the combination / interaction of the categories}
}
\value{
A matrix representing the categorical variables in binary form ("dummy coding")
}
\description{
Get binary representation of categorical variables
}
\details{
This function can be used to include categorical variables as part of the 
optimization criterion in k-means / k-plus anticlustering, rather than 
including them as hard constraints as done in \code{\link{anticlustering}}.
See examples.
}
\examples{

# Use Schaper data set for example
data(schaper2019)
features <- schaper2019[, 3:6]
K <- 3
N <- nrow(features) 

input_data <- cbind(
  kplus_moment_variables(features, T = 2), 
  categories_to_binary(schaper2019$room)
)

kplus_groups <- anticlustering(
  input_data, 
  K = K,
  objective = "variance",
  method = "local-maximum", 
  repetitions = 10
)
mean_sd_tab(features, kplus_groups)
table(kplus_groups, schaper2019$room) # categories argument was not used


}
\references{
Papenberg, M. (in press). k-plus Anticlustering: An Improved k-means 
Criterion for Maximizing Between-Group Similarity. British Journal of 
Mathematical and Statistical Psychology.
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
