% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering}
\alias{clustering}
\title{Create clusters of equal size}
\usage{
clustering(features, n_clusters, standardize = TRUE, method = "heuristic")
}
\arguments{
\item{features}{A vector, matrix or data.frame of data points. Rows
correspond to elements and columns correspond to features.}

\item{n_clusters}{How many anticlusters should be created.}

\item{standardize}{Boolean - should the features be standardized
before anticlusters are assigned? Defaults to TRUE}

\item{method}{One of "heuristic" or "exact". When "exact" is chosen,
the function minimizes the distance criterion; when "heuristic"
is chosen, the function tries to minimize the variance criterion
(see details). Use method exact only for rather small problem
sizes (< xyz elements).}
}
\value{
A vector representing the cluster affiliation of all elements.
}
\description{
Create clusters of equal size
}
\details{
TODO: Write details
}
\examples{

# TODO: Write examples

}
