% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted-cluster-editing.R
\name{cpp_similarities}
\alias{cpp_similarities}
\title{Compute similarity values for the clique partitioning problem}
\usage{
cpp_similarities(x)
}
\arguments{
\item{x}{A data frame / matrix of categorical (nomial / ordinal) attributes}
}
\value{
An object of class \code{dist} representing pairwise similarity (not (DIS)similartiy!)
}
\description{
Compute similarity values for the clique partitioning problem
}
\details{
"the [...] values represent the number of attributes on which vertices
[...] disagree, minus the number of attributes on which they agree." 
(Brusco et al. 2009, p. 689)
}
\references{
Brusco, M. J., & Köhn, H. F. (2009). Clustering qualitative data
based on binary equivalence relations: neighborhood search
heuristics for the clique partitioning problem. Psychometrika, 74,
685–703.

Grötschel, M., & Wakabayashi, Y. (1989). A cutting plane algorithm
for a clustering problem. Mathematical Programming, 45, 59-96.
}
