% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective-anticluster-editing.R
\name{distance_objective}
\alias{distance_objective}
\title{Cluster editing distance objective}
\usage{
distance_objective(x, clusters)
}
\arguments{
\item{x}{The data input. Can be one of two structures: (1) A data matrix
where rows correspond to elements and columns correspond to
features (a single numeric feature can be passed as a vector). (2)
An N x N matrix dissimilarity matrix; can be an object of class
\code{dist} (e.g., returned by \code{\link{dist}} or
\code{\link{as.dist}}) or a \code{matrix} where the entries of
the upper and lower triangular matrix represent the pairwise
dissimilarities.}

\item{clusters}{A vector representing (anti)clusters (e.g.,
returned by \code{\link{anticlustering}}).}
}
\value{
The cluster editing objective
}
\description{
Compute the cluster editing objective for a given clustering.
}
\details{
The cluster editing objective objective is given by the sum of the
pairwise distances between elements within the same (anti)clusters.
When the input \code{x} is a feature matrix, the Euclidean distance
is computed as the basic distance unit.
}
\examples{

data(iris)
distances <- dist(iris[1:60, -5])
## Clustering
clusters <- balanced_clustering(distances, K = 3)
# This is low:
distance_objective(distances, clusters)
## Anticlustering
anticlusters <- anticlustering(distances, K = 3)
# This is higher:
distance_objective(distances, anticlusters)

}
\references{
Papenberg, M., & Klau, G. W. (2019, October 30). Using anticlustering to partition 
data sets into equivalent parts https://doi.org/10.31234/osf.io/3razc
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
