% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilp-postprocessing.R
\name{edit_distances}
\alias{edit_distances}
\title{Edit distances}
\usage{
edit_distances(distances, clustering, value = -1e+06)
}
\arguments{
\item{distances}{A distance object or matrix of
between-item-distances.}

\item{clustering}{A vector representing a clustering; objects that
are part of the same cluster are assigned a new distance.}

\item{value}{The value that is assigned to the fixed distances.
Defaults to -1,000,000.}
}
\value{
A distance object containing the fixed distances. For items
    that are part of the same cluster.
}
\description{
Based on a clustering, distances between elements of the same cluster
are set to a specified value. This is used to enforce or forbid pairs
of elements to be part of the same anticluster when the distance
matrix is passed to the ILP solver.  By considering a preclustering
of items, very similar items will not be assigned to the same group
when the fixed distance object is used to create the ILP formulation
of the item assignment instance.
}
