% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete-enumeration.R
\name{enum_anticlustering}
\alias{enum_anticlustering}
\title{Complete enumeration approach to anticluster editing}
\usage{
enum_anticlustering(features = NULL, distances = NULL, K)
}
\arguments{
\item{features}{A vector, matrix or data.frame of data points. Rows
correspond to elements and columns correspond to features. A
vector represents a single feature.}

\item{distances}{Alternative data argument that can be used if
\code{features} is not used. A N x N matrix representing the
pairwise dissimilarities between all N elements. Larger values
indicate higher dissimilarity. Can be an object of class
\code{dist} (e.g., returned by \code{\link{dist}} or
\code{\link{as.dist}}.}

\item{K}{How many anticlusters should be created.}
}
\value{
The anticluster affiliation
}
\description{
Computes the optimum anticluster editing objective but is probably
very slow because all possible partitions of anticlusters are
enumerated.
}
\examples{

n_features <- 2
N <- 10
K <- 2
features <- matrix(runif(n_features * N), ncol = n_features)
results <- enum_anticlustering(features, K = K)
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
