% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{equal_sized_cluster_editing}
\alias{equal_sized_cluster_editing}
\title{Solve exact equal-sized cluster editing}
\usage{
equal_sized_cluster_editing(items, n_groups, solver, standardize = FALSE)
}
\arguments{
\item{items}{A data.frame of item features. Rows must correspond to
items and columns to features.}

\item{n_groups}{How many clusters are to be created.}

\item{solver}{A string identifing the solver to be used ("Rglpk",
"gurobi", or "cplex")}

\item{standardize}{Boolean - should the feature values be
standardized before groups are created? Defaults to FALSE.}
}
\value{
A vector representing the clustering.
}
\description{
Solve exact equal-sized cluster editing
}
