% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heuristic.R
\name{equal_sized_clustering}
\alias{equal_sized_clustering}
\title{Heuristic clustering algorithm creating equal-sized clusters}
\usage{
equal_sized_clustering(items, nclusters)
}
\arguments{
\item{items}{A vector, matrix or data.frame of data points.
Rows correspond to items and columns correspond to features.}

\item{nclusters}{The number of clusters to be created. Must be a
factor of the number of items.}
}
\value{
A vector representing the clustering
}
\description{
Uses kmeans algorithm to initiate cluster centers and then
sequentially chooses a cluster center and assigns assigns the closest
item to it. This ensures that each cluster is filled with the same
number of items.
}
