% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c-anticlustering.R
\name{fanticlust}
\alias{fanticlust}
\title{C implementation of anticlustering}
\usage{
fanticlust(data, clusters, categories = NULL)
}
\arguments{
\item{data}{A N x M data matrix}

\item{clusters}{An initial assignment of the elements to clusters}

\item{categories}{A vector, data.frame or matrix representing one
or several categorical constraints.}
}
\description{
C implementation of anticlustering
}
\examples{
# Compare classical anticlustering implementation and C implementation
N <- 100
M <- 3
data <- matrix(rnorm(N * M), ncol = M)
K <- 2
n_cats <- 2
categories <- sample(n_cats, size = N, replace = TRUE)
clusters <- anticlust:::categorical_sampling(categories, K)
start <- Sys.time()
cl1 <- fanticlust(data, clusters, categories)
Sys.time() - start
table(categories, cl1)

start <- Sys.time()
cl2 <- anticlustering(
  data,
  K = clusters,
  objective = "variance"
)
Sys.time() - start

all(cl1 == cl2) # Same algorithm in C and R


}
