% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c-anticlustering.R
\name{fanticlust}
\alias{fanticlust}
\title{C implementation of anticlustering}
\usage{
fanticlust(data, clusters)
}
\arguments{
\item{data}{A N x M data matrix}

\item{clusters}{An initial assignment of the elements to clusters}
}
\description{
C implementation of anticlustering
}
\examples{
# Compare classical anticlustering implementation and C implementation
N <- 200
M <- 3
data <- matrix(rnorm(N * M), ncol = M)
K <- 3
clusters <- sample(rep_len(1:K, N))
start <- Sys.time()
cl1 <- fanticlust(data, clusters)
Sys.time() - start

start <- Sys.time()
cl2 <- anticlustering(
  data,
  K = clusters,
  objective = "variance"
)
Sys.time() - start

all(cl1 == cl2) # Same algorithm in C and R


}
