% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange-method-kmeans-anticlustering.R
\name{fast_anticlustering}
\alias{fast_anticlustering}
\title{Fast anticlustering}
\usage{
fast_anticlustering(x, K, k_neighbours = Inf, categories = NULL, backend = "C")
}
\arguments{
\item{x}{A numeric vector, matrix or data.frame of data
points.  Rows correspond to elements and columns correspond to
features. A vector represents a single numeric feature.}

\item{K}{How many anticlusters should be created. Alternatively:
(a) A vector describing the size of each group, or (b) a vector
of length \code{nrow(x)} describing how elements are assigned
to anticlusters before the optimization starts.}

\item{k_neighbours}{The number of neighbours that serve as exchange
partner for each element. Defaults to Inf, implying that each element
is exchanged with each element in other groups.}

\item{categories}{A vector, data.frame or matrix representing one or
several categorical constraints.}

\item{backend}{Either "C" or "R", to use a C or R implementation of the
anticlustering optimization algorithm. Since \code{anticlust} version 
0.7.1, the faster C implementation is the default.}
}
\description{
Anticlustering via optimizing the k-means criterion with an
adjusted exchange method where the number of exchange partners can be 
specified. Using fewer exchange partners can speed up the computation 
and make anticlustering applicable to quite large data sets.
}
\details{
This function was created to make anticlustering applicable
to large data sets (e.g., several 100,000 elements). It optimizes the k-means
objective because computing all pairwise as is done when optimizing 
the diversity is not feasible for very large data sets (for about N > 20000 on my 
personal computer). 
Additionally, this function employs a
speed-optimized exchange method by (possibly) not using each element as exchange partner. 
Instead, for each element, the potential
exchange partners are generated using a nearest neighbor search with the
function \code{\link[RANN]{nn2}} from the \code{RANN} package. The nearest
neighbors then serve as exchange partners. This approach is inspired by the
preclustering heuristic according to which good solutions are found
when similar elements are in different sets---by swapping nearest
neighbors, this will often be the case. The number of exchange partners
per element has to be set using the argument \code{k_neighbours}; by
default, it is set to \code{Inf}, meaning that all possible swaps are
tested. This default must be changed by the user for large data sets.
More exchange partners generally improve the output, but also increase
run time.

When setting the \code{categories} argument, exchange partners will
be generated from the same category. Note that when
\code{categories} has multiple columns (i.e., each element is
assigned to multiple columns), each combination of categories is
treated as a distinct category by the exchange method. You can use 
\code{\link{categories_to_binary}} to improve results for several categorical
variables.
}
\examples{

features <- iris[, - 5]

start <- Sys.time()
ac_exchange <- fast_anticlustering(features, K = 3)
Sys.time() - start

## The following call is equivalent to the call above:
start <- Sys.time()
ac_exchange <- anticlustering(features, K = 3, objective = "variance")
Sys.time() - start

## Improve run time by using fewer exchange partners:
start <- Sys.time()
ac_fast <- fast_anticlustering(features, K = 3, k_neighbours = 10)
Sys.time() - start

# Use k-plus anticlustering with this function is straight forward,
# just use kplus_moment_variables() on the numeric input:
kplus_anticlusters <- fast_anticlustering(kplus_moment_variables(features, T = 2), K = 3)
mean_sd_tab(features, kplus_anticlusters)

# Working on several 1000 elements is incredibly fast (Here n = 5000)
data <- matrix(rnorm(5000 * 2), ncol = 2)
groups <- fast_anticlustering(data, K = 2, k_neighbours = 2)
mean_sd_tab(data, groups)

}
\seealso{
\code{\link{anticlustering}}

\code{\link{kplus_moment_variables}}

\code{\link{variance_objective}}
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
