% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast-exchange-method.R
\name{fast_anticlustering}
\alias{fast_anticlustering}
\title{Fast anticlustering}
\usage{
fast_anticlustering(features, K, k_neighbours = Inf, categories = NULL)
}
\arguments{
\item{features}{A numeric vector, matrix or data.frame of data
points.  Rows correspond to elements and columns correspond to
features. A vector represents a single numeric feature.}

\item{K}{How many anticlusters should be created.}

\item{k_neighbours}{The number of neighbours that serve as exchange
partner for each element. Defaults to Inf, i.e., each element
is exchanged with each element in other groups.}

\item{categories}{A vector, data.frame or matrix representing one or
several categorical constraints.}
}
\description{
The most efficient way to solve anticlustering optimizing the
k-means variance criterion. Can be used for very large data sets.
}
\examples{

start <- Sys.time()
features <- iris[, - 5]
ac_exchange <- fast_anticlustering(features, K = 3)
Sys.time() - start
by(features, ac_exchange, function(x) round(colMeans(x, na.rm = TRUE), 2))

start <- Sys.time()
features <- schaper2019[, 3:6]
ac_exchange <- fast_anticlustering(features, K = 3, categories = schaper2019$room)
Sys.time() - start
by(features, ac_exchange, function(x) round(colMeans(x, na.rm = TRUE), 2))

}
