% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heuristic.R
\name{heuristic_anticlustering}
\alias{heuristic_anticlustering}
\title{Anticlustering based on a heuristic}
\usage{
heuristic_anticlustering(features, clustering, nrep = 100)
}
\arguments{
\item{features}{A data.frame, matrix or vector representing the features
that are used.}

\item{clustering}{A vector representing the preclustering of elements.
See Details.}
}
\value{
A vector representing the anticlustering.
}
\description{
Anticlustering based on a heuristic
}
\details{
The heuristic approach to anticlustering forbids elements that are
part of the same
precluster to be assigned to the same group. The preclustering should be accomplished by one of the clustering
functions, `equal_sized_cluster_editing` (an exact method that minimizes
distance criterion under the restriction of equal group sizes) or
`equal_sized_clustering` (a heuristic method that tries to minimize
the variance criterion under the restriction of equal group sizes).
}
