% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heuristic.R
\name{heuristic_item_assignment}
\alias{heuristic_item_assignment}
\title{Based on a preclustering, this function creates a heuristic item assignment}
\usage{
heuristic_item_assignment(assignment, nrep = 100)
}
\arguments{
\item{assignment}{A data.frame that represents the preclustering assingment.
Usually, the return value of `ilp_to_groups` or `equal_sized_clustering`
will be the input.}
}
\value{
The data.frame assigment that was passed, but now the column
  `group` contains the groups as determined for the item assignment
  problem.

  @export
}
\description{
Based on a preclustering, this function creates a heuristic item assignment
}
