% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heuristic.R
\name{heuristic_item_assignment}
\alias{heuristic_item_assignment}
\title{Based on a preclustering, create a heuristic item assignment}
\usage{
heuristic_item_assignment(features, clustering, nrep = 100)
}
\arguments{
\item{features}{A data.frame representing the features that are used
in the assignment.}

\item{clustering}{A vector representing the clustering of elements}
}
\value{
A vector representing the new group assignment.
}
\description{
Based on a preclustering, create a heuristic item assignment
}
