% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILP_TO_GROUP.R
\name{ilp_to_groups}
\alias{ilp_to_groups}
\title{Return group membership based on ILP solution for item assignment}
\usage{
ilp_to_groups(ilp, solution)
}
\arguments{
\item{ilp}{An ILP representation of the item assignment problem
(returned by rom `item_assign_ilp`)}

\item{solution}{The solution of the instance returned by the function
`solve_ilp`.}
}
\value{
A data.frame containing one column of item ids (here, the id
    corresponds to the order of the items) and one column contains the
    group assignments of the items. The original items are also
    returned as columns of the data.frame.
}
\description{
This function is primarily called from within other functions and probably
should not be called directly.
}
