% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILP.R
\name{item_assign_ilp}
\alias{item_assign_ilp}
\title{Construct the ILP represenation of an item assignment problem}
\usage{
item_assign_ilp(distances, p, solver = "Rglpk")
}
\arguments{
\item{distances}{An distance object or matrix representing the
distances between items}

\item{p}{The number of groups to be created}

\item{solver}{A string identifing the solver to be used ("Rglpk",
"gurobi", or "cplex")}
}
\value{
A list representing ILP formulation of the instance
}
\description{
Construct the ILP represenation of an item assignment problem
}
