% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching-wrapper.R
\name{matching}
\alias{matching}
\title{Matching}
\usage{
matching(features = NULL, distances = NULL, p = 2, groups = NULL)
}
\arguments{
\item{features}{A numeric vector, matrix or data.frame of data
points.  Rows correspond to elements and columns correspond to
features. A vector represents a single numeric feature.}

\item{distances}{Alternative data argument that can be used if
\code{features} is not passed. An N x N matrix representing the
pairwise dissimilarities between N elements. Larger values
indicate higher dissimilarity. Can be an object of class
\code{dist} (e.g., returned by \code{\link{dist}} or
\code{\link{as.dist}}) or a \code{matrix} where the entries of
the upper and lower triangular matrix represent the pairwise
dissimilarities.}

\item{p}{The size of the groups; the default is 2, in which case the
function returns pairs.}

\item{groups}{An optional vector inducing grouping restrictions. If the
vector is passed, the argument \code{p} is ignored and matches are
sought between elements of different groups.}
}
\value{
An integer vector encoding the matches. See details for more information.
}
\description{
Conduct K-partite or unrestricted (minimum distance) matching to find pairs 
or groups of similar elements. Finding matches is by default based on the Euclidean 
distance between data points, but a custom distance measure can also be employed.
}
\details{
If the argument \code{features} is passed, matching is done based on the Euclidean
distance between data points. To find matches, this function uses
the same algorithm as implemented in \code{\link{balanced_clustering}}
but it differs with regard to the interface: This function specifies the size of the groups,
\code{\link{balanced_clustering}} specifies 
the number of the clusters. Moreover, this function makes it possible to specify grouping 
restrictions using the \code{groups} argument, thus enabling K-partite minimum matching.

The output of this function is an integer vector encoding which elements have 
been matched. The grouping numbers are sorted by similarity. That is, elements with the
grouping number »1« are most similar to each other, followed by 2 etc (groups having different
similarity indices are still assigned a different grouping number). Similarity is measured
as the sum of pairwise (Euclidean) distances within groups \code{\link{distance_objective}}.
Some elements of the output may be \code{NA}. This happens if it is not possible to evenly
split the item pool evenly into groups of size \code{p} or if the categories described by
the argument \code{groups} are of different size. Unmatched items are then assigned \code{NA}.
}
\examples{

# Find triplets
N <- 299 # two elements will not be matched
lds <- data.frame(f1 = rnorm(N), f2 = rnorm(N))
triplets <- matching(lds, p = 3)
table(triplets)
sum(is.na(triplets))
plot_clusters(lds, clustering = triplets, within_connection = TRUE)

# Match between different plant species
species <- iris$Species != "versicolor"
matched <- matching(iris[species, 1], groups = iris[species, 5])
plot_clusters(
  data.frame(Species = as.numeric(iris[species, 5]), Sepal.Length = iris[species, 1]),
  clustering = matched,
  within_connection = TRUE
)

# Use unequal-sized groups: Only selects matches for some elements
N <- 33
data <- matrix(rnorm(N), ncol = 1)
groups <- sample(1:2, size = N, replace = TRUE, prob = c(0.7, 0.3))
matched <- matching(data[, 1], groups = groups)
plot_clusters(
  cbind(groups, data), 
  clustering = matched, 
  within_connection = TRUE
)


}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
