% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective-mean-sd.R
\name{mean_sd_obj}
\alias{mean_sd_obj}
\title{An objective function measuring similarity of sets}
\usage{
mean_sd_obj(clusters, features)
}
\arguments{
\item{clusters}{A clustering vector}

\item{features}{The features}
}
\value{
A value quantifying set similarity. Higher values indicate
    that the discrepancy in means and standard deviations
    are low.
}
\description{
Defines set similarity as the discrepancy in means and standard deviations
}
\details{
Use this function as the \code{objective} in the anticlustering
function to make means and standard deviations as similar
as possible.
}
\examples{
data(schaper2019)
features <- schaper2019[1:48, 3:6]
ac <- anticlustering(
  features,
  K = 3,
  categories = schaper2019$room[1:48],
  objective = mean_sd_obj
)
by(features, ac, function(x) round(colMeans(x), 2))
by(features, ac, function(x) round(apply(x, 2, sd), 2))

}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
