% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_sd_objective-function.R
\name{mean_sd_obj}
\alias{mean_sd_obj}
\title{An objective function measuring similarity via discrepancy features}
\usage{
mean_sd_obj(clusters, features)
}
\arguments{
\item{clusters}{A clustering vector}

\item{features}{The features}
}
\value{
A value quantifying set similarity as the total absolute deviations
    between sets on all features.
}
\description{
Uses means and standard deviations
}
