% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/must-link-anticlustering.R
\name{must_link_anticlustering}
\alias{must_link_anticlustering}
\title{Must-link anticlustering}
\usage{
must_link_anticlustering(x, K, must_link, infeasible_prob = 0)
}
\arguments{
\item{x}{A distance matrix (Only maximum diversity is implemented)}

\item{K}{What do the required groups look like (as in \code{\link{anticlustering}})}

\item{must_link}{Vector of IDs; elements having the same value are linked in the same anticluster.}

\item{infeasible_prob}{Probability that a swap is accepted that does not maintain the required balance in group sizes. Defaults to 0.}
}
\description{
Must-link anticlustering
}
\details{
Conducts an exchange method (method = "local-maximum") on "merged" elements.
}
