% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/must-link-anticlustering.R
\name{must_link_anticlustering}
\alias{must_link_anticlustering}
\title{This is the function that is called from anticlustering()}
\usage{
must_link_anticlustering(
  x,
  K,
  must_link,
  method = "exchange",
  objective = "diversity",
  repetitions = NULL
)
}
\description{
This is the function that is called from anticlustering()
}
