% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heuristic.R
\name{obj_value}
\alias{obj_value}
\title{Get the objective function from an assignment object}
\usage{
obj_value(assignment, features)
}
\arguments{
\item{assignment}{A vector representing the group assignment}

\item{features}{A data.frame representing the features that were used
in the assignment.}
}
\value{
Scalar: the objective value.
}
\description{
Get the objective function from an assignment object
}
