% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_anticlustering.R
\name{optimal_anticlustering}
\alias{optimal_anticlustering}
\title{Optimal ("exact") algorithms for anticlustering}
\usage{
optimal_anticlustering(x, K, objective, solver = NULL)
}
\arguments{
\item{x}{The data input. Can be one of two structures: (1) A
feature matrix where rows correspond to elements and columns
correspond to variables (a single numeric variable can be
passed as a vector). (2) An N x N matrix dissimilarity matrix;
can be an object of class \code{dist} (e.g., returned by
\code{\link{dist}} or \code{\link{as.dist}}) or a \code{matrix}
where the entries of the upper and lower triangular matrix
represent pairwise dissimilarities.}

\item{K}{How many anticlusters should be created or alternatively:
(a) A vector describing the size of each group (the latter currently 
only works for \code{objective = "dispersion")}.}

\item{objective}{The anticlustering objective, can be "diversity", "variance", 
"kplus" or "dispersion".}

\item{solver}{Optional. The solver used to obtain the optimal method. 
Currently supports "glpk" and "symphony". See details.}
}
\value{
A vector of length N that assigns a group (i.e, a number
    between 1 and \code{K}) to each input element.
}
\description{
Wrapper function that gives access to all optimal algorithms for anticlustering 
that are available in anticlust.
}
\details{
This is a wrapper for all optimal methods supported in anticlust. As compared to 
\code{\link{anticlust}}, it allows to specify the solver to obtain an optimal
solution and it can be used to obtain optimal solutions for all supported
anticlustering objectives (variance, diversity, k-plus, dispersion). For 
the objectives variance, diversity and k-plus, the optimal ILP method 
in Papenberg and Klau (2021) is used. For the variance (i.e., k-means anticlustering), 
the squared Euclidean distance is used, and for k-plus the squared Euclidean distance
based on the extended k-plus data matrix is used (see \code{\link{kplus_moment_variables}}).
For the diversity (and the dispersion), the Euclidean distance is used by default, 
but any user-defined dissimilarity matrix is possible.

The dispersion is solved optimal using the approach described in \code{\link{optimal_dispersion}}.

The optimal methods either require the R package \code{Rglpk} and the GNU linear programming kit
(<http://www.gnu.org/software/glpk/>), or the R package
\code{Rsymphony} and the COIN-OR SYMPHONY solver libraries
(<https://github.com/coin-or/SYMPHONY>). If the argument \code{solver} is not 
specified by the user, the function will try to find the GLPK or SYMPHONY 
solver and throw an error if none is available. It will usually select the 
SYMPHONY solver if both are available because it usually performs better than 
the GLPK solver. For Mac computers, however, the GLPK is selected by default if both solvers 
are available. This is because some rare instances have been observed where
the SYMPHONY solver crashes on Mac computers. I would still try out the 
SYMPHONY solver to see if the unlikely crash occurs. However, this has to be 
set by the user.
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
