% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimal_Dispersion.R
\name{optimal_dispersion}
\alias{optimal_dispersion}
\title{Maximize dispersion for K groups}
\usage{
optimal_dispersion(distances, K, solver = "glpk")
}
\arguments{
\item{distances}{Input data, dissimilarity matrix}

\item{K}{The number of anticlusters to be created}

\item{solver}{Either "glpk" (default) or "symphony"}
}
\value{
A list with three elements:  
  $dispersion: The optimal dispersion; 
  $groups: A possible assignment to groups (vector);
  $edges: A matrix of 2 columns. Each row is an edge that had to be investigated to find the dispersion.
}
\description{
Maximize dispersion for K groups
}
