% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schaper2019.R
\docType{data}
\name{schaper2019}
\alias{schaper2019}
\title{Ratings for 96 objects}
\format{A data frame with 96 rows and 7 variables
\describe{
  \item{item}{The name of the object (in German)}
  \item{room}{The room in which the item is typically found; can be
      'kitchen' or 'bathroom'}
  \item{rating_consistent}{How expected would it
      be to find the \code{item} in the typical \code{room}}
  \item{rating_inconsistent}{How expected would it
      be to find the \code{item} in the atypical \code{room}}
  \item{syllables}{The number of syllables in the object name}
  \item{frequency}{A value indicating the relative frequency of the
      object name in German language}
  \item{list}{Represents the set affiliation of the \code{item} as
      realized in experiments by Schaper et al.}
}}
\source{
Courteously provided by Marie Lusia Schaper and Ute Bayen.
}
\usage{
schaper2019
}
\description{
An item set that was used in experiments by Schaper, Kuhlmann and
Bayen (2019; in press).
}
\examples{

features <- schaper2019[, 3:6]
head(features)
# This command scales all features to have a similar range:
features <- apply(features, 2, function(x) x / max(x))
head(features)


# Optimize the variance criterion
# (tends to maximize similarity in feature means)
anticlusters <- anticlustering(
  features,
  K = 3,
  objective = "variance",
  categories = schaper2019$room,
  seed = 625,
  nrep = 1000 # increase for better results
)

# Means are quite similar across sets:
by(schaper2019[, 3:6], anticlusters, function(x) round(colMeans(x), 2))
# Room is balanced between the three sets:
table(Room = schaper2019$room, Set = anticlusters)

}
\references{
Schaper, M. L., Kuhlmann, B. G., & Bayen, U. J. (2019). Metamemory
expectancy illusion and schema-consistent guessing in source monitoring.
Journal of Experimental Psychology: Learning, Memory, and Cognition,
45, 470-496. http://dx.doi.org/10.1037/xlm0000602

Schaper, M. L., Kuhlmann, B. G., & Bayen, U. J. (in press).
Metacognitive expectancy effects in source monitoring: Beliefs,
in-the-moment experiences, or both? Journal of Memory and Language.
}
\keyword{datasets}
