% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_select.R
\name{select_stimuli}
\alias{select_stimuli}
\title{Select stimuli for experiments}
\usage{
select_stimuli(data, split_by = NULL, equalize, balance = NULL, design,
  n = NULL)
}
\arguments{
\item{data}{A N x M data frame of features describing stimuli}

\item{split_by}{Character vector, the names of the variables in 
\code{data} that should be different between sets}

\item{equalize}{Character vector, the names of the variables in 
\code{data} that should be similar across sets}

\item{balance}{One or more categorical variables. The frequency of
the categories is balanced between sets.}

\item{design}{Specifies the number of groups per \code{split_by} 
feature. Is a vector of length \code{ncol(split_by)} (or of length
1 if only one - or no - \code{split_by} feature is passed).}

\item{n}{The number of elements per set.}

\item{p}{The number of exchange partners per stimulus used by the
exchange method that optimizes similarity with regard to the 
\code{equalize} variables. Higher values increase
the precision of the results but also increase run time.}
}
\value{
A data frame that has the same columns as the original input 
   (the data frame \code{data}), but has an additional called \code{SET}.
}
\description{
Stimulus selection via `Divide and Select` or `anticlustering`.
}
\details{
The argument \code{split_by} will convert a numeric variable into 
a categorical variable, and will recognize if a variable is already
categorical by testing if \code{length(unique(split_by)) == design}
is \code{TRUE}.
}
\examples{
# TODO

}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
