% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_stimuli.R
\name{select_stimuli}
\alias{select_stimuli}
\title{Select stimuli for experiments}
\usage{
select_stimuli(data, split_by = NULL, equalize, balance = NULL, design,
  n = NULL)
}
\arguments{
\item{data}{A N x M data frame of features describing stimuli}

\item{split_by}{Character vector, the names of the variables in 
\code{data} that should be different between sets}

\item{equalize}{Character vector, the names of the variables in 
\code{data} that should be similar across sets}

\item{balance}{Character vector, the names of the variables in 
\code{data} that constitute categorical variables whose frequency
should be balanced across sets}

\item{design}{Specifies the number of groups per \code{split_by} 
feature. Is a vector of length \code{ncol(split_by)} (or of length
1 if one or no \code{split_by} feature is passed).}

\item{n}{The number of elements per set. If not specified, the complete 
date set is partitioned into subsets (the number of subsets is specified
using the \code{design} argument).}
}
\value{
A data frame that has the same columns as the original input 
   (the data frame \code{data}), but has an additional called \code{SET}.
}
\description{
Stimulus selection via Anticlustering or Min-Max-Anticlustering
}
\examples{
# TODO

}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
