% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_stimuli.R
\name{select_stimuli}
\alias{select_stimuli}
\title{Select stimuli for experiments}
\usage{
select_stimuli(data, split_by = NULL, equalize, design, n = NULL,
  randomness = 1)
}
\arguments{
\item{data}{A N x M data frame of features describing stimuli}

\item{split_by}{Character vector, the names of the variables in 
\code{data} that should be different between sets}

\item{equalize}{Character vector, the names of the variables in 
\code{data} that should be similar across sets}

\item{design}{Specifies the number of groups per \code{split_by} 
feature. Is a vector of length \code{ncol(split_by)} (or of length
1 if only one - or no - \code{split_by} feature is passed).}

\item{n}{The number of elements per set.}

\item{randomness}{An integer in `1:3`. Used for subset anticlustering.
1 = the same stimuli are returned each call; 2 = there is some 
randomness in which stimuli are returned. 3 = there is more 
randomness in which stimuli are returned.}
}
\value{
A data frame that has the same columns as the original input 
   (the data frame \code{data}), but has an additional called \code{SET}.
}
\description{
Stimulus selection via Anticlustering or Min-Max-Anticlustering
}
\examples{
# TODO

}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
