% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-sort.R
\name{sort_by_col}
\alias{sort_by_col}
\title{Sort a matrix or data frame by one column}
\usage{
sort_by_col(x, col)
}
\arguments{
\item{x}{A matrix or data frame}

\item{col}{An numeric index or character vector of length 1 to select
the column by which the table is sorted.}
}
\value{
The matrix or data frame sorted by `row`
}
\description{
Sort a matrix or data frame by one column
}
