% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/wrapper-three-phase-search-dynamic-population-size.R
\name{three_phase_search_anticlustering}
\alias{three_phase_search_anticlustering}
\title{Three phase search with dynamic population size heuristic}
\usage{
three_phase_search_anticlustering(
  matrix,
  K,
  N,
  upper_bound = NULL,
  lower_bound = NULL,
  popSize = 15,
  time_limit = NULL,
  theta_max = NULL,
  theta_min = NULL,
  beta_min = NULL,
  LMAX = 3
)
}
\arguments{
\item{matrix}{The data input. Currently just a vector.}

\item{K}{How many anticlusters should be created.}

\item{popSize}{beta_max from paper}

\item{return}{}
}
\description{
This function implements the three phase search algorithm TPSPD for
anticlustering by Yang et al. (2022; <doi.org/10.1016/j.ejor.2022.02.003>).
The description of their algorithm is
given in Section 2 of their paper (in particular, see the
Pseudocode in Algorithm 1).
}
\references{
Xiao Yang et al. “A three-phase search approach with dynamic population size for solving # nolint
the maximally diverse grouping problem”. In: European Journal of Operational Research
302.3 (2022). [SOURCE-CODE: https://raw.githubusercontent.com/toyamaailab/toyamaailab.github.io/main/resource/TPSDP_Code.zip],
 pp. 925–953. ISSN: 0377-2217. DOI: https://doi.org/10.1016/j.ejor.2022.02.003. 

#' special imports for rccp and eval function
}
\author{
Hannah Hengelbrock \email{Hannah.Hengelbrock@hhu.de}, 
Martin Papenberg \email{martin.papenberg@hhu.de}
}
