% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted-cluster-editing.R
\name{wce}
\alias{wce}
\title{Solve weighted cluster editing exactly}
\usage{
wce(weights)
}
\arguments{
\item{weights}{A N x N weight matrix. Larger values indicate stronger
agreement (this is unlike in the \code{\link{anticlustering}} function)}
}
\value{
The clusters
}
\description{
Solve weighted cluster editing exactly
}
\examples{
subs <- sample(150, size = 60)
data <- iris[subs, -5]
distances <- dist(data)
# Define agreement as being close enough to each other
agreements <- ifelse(as.matrix(distances) < 2.5, 1, -1)
clusters <- wce(agreements)
plot_clusters(data[, 1:2], clusters)

}
