% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted-cluster-editing.R
\name{wce}
\alias{wce}
\title{Exact weighted cluster editing}
\usage{
wce(weights)
}
\arguments{
\item{weights}{A N x N weight matrix. Larger values indicate stronger
agreement (this is unlike in the \code{\link{anticlustering}} function)}
}
\value{
An integer vector representing the cluster affiliation of each data point
}
\description{
Exact weighted cluster editing
}
\details{
To obtain an optimal solution for weighted cluster editing, 
a linear programming solver must be installed
and usable from R. The \code{anticlust} package 
requires the open source GNU linear programming kit, which 
is called from the package \code{Rglpk}).
}
\examples{
\dontrun{
features <- swiss
distances <- dist(scale(swiss))
hist(distances)
# Define agreement as being close enough to each other.
# By defining low agreement as -1 and high agreement as +1, we
# solve *unweighted* cluster editing
agreements <- ifelse(as.matrix(distances) < 3, 1, -1)
clusters <- wce(agreements)
plot(swiss, col = clusters, pch = 19)
}

}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
